basic,diagnostic: move diagnostic related to diagnostics.go
