+++ diagnostics_gojay_test.go
@@ -12,6 +12,14 @@ import (
 	"github.com/francoispqt/gojay"
 )
 
+func TestDiagnostic(t *testing.T) {
+	testDiagnostic(t, gojay.Marshal, gojay.Unsafe.Unmarshal)
+}
+
+func TestDiagnosticRelatedInformation(t *testing.T) {
+	testDiagnosticRelatedInformation(t, gojay.Marshal, gojay.Unsafe.Unmarshal)
+}
+
 func TestPublishDiagnosticsParams(t *testing.T) {
 	testPublishDiagnosticsParams(t, gojay.Marshal, gojay.Unsafe.Unmarshal)
 }
